requirements.txt
psycopg2==2.8.3
Django==2.2.4
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
requests
gunicorn